Support for uploading TBv2 collected data to a distinct key.
- Upload statistics and user recordings from TBv2 devices to
  'collected-data.v2'.
-- The data are a different format, and this provides a simple and reliable
   means to distinguish them.

+++ acm/src/main/java/org/literacybridge/acm/tbloader/StatisticsUploader.java
@@ -5,7 +5,7 @@ import org.literacybridge.acm.cloud.Authenticator;
 import org.literacybridge.acm.config.ACMConfiguration;
 import org.literacybridge.core.fs.ZipUnzip;
 
-import javax.swing.*;
+import javax.swing.SwingWorker;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
@@ -17,43 +17,41 @@ import java.util.List;
 public class StatisticsUploader {
     private final TBLoader tbLoader;
 
-    private final File collectionWorkDir;
     private final File uploadQueueDir;
-    private final File uploadQueueCDDir;
-    private final File uploadQueueDeviceDir;
 
-    private final boolean suppressUpload = ACMConfiguration.getInstance().isSuppressStatisticsUpload();
+    private boolean suppressUpload = ACMConfiguration.getInstance().isSuppressStatisticsUpload();
 
     StatisticsUploader(TBLoader tbLoader,
-        File collectionWorkDir,
-        File uploadQueueDir,
-        File uploadQueueCDDir,
-        File uploadQueueDeviceDir) {
+        File uploadQueueDir) {
 
         this.tbLoader = tbLoader;
-        this.collectionWorkDir = collectionWorkDir;
         this.uploadQueueDir = uploadQueueDir;
-        this.uploadQueueCDDir = uploadQueueCDDir;
-        this.uploadQueueDeviceDir = uploadQueueDeviceDir;
     }
 
     /**
-     * Look for directories in collectionWorkDir. For any that are found, zip their contents into
-     * a single file, and move it to the upload queue. Then, if the upload worker has not been
-     * started, start it.
+     * Enqueue for upload any files in fromDirectory. Any sub-directories in fromDirectory
+     * will be zipped to a file with the sub-directory name + ".zip", then enqueued.
+     *
+     * @param fromDirectory A directory from which to enqueue files.
+     * @param keyPrefix     A prefix to prepend the file name to construct the S3 key, for
+     *                      example "collected-data.v2/tbcd000c"
      */
-    void zipAndUpload() {
-        File[] uploadables = collectionWorkDir.listFiles();
+    void zipAndEnqueue(File fromDirectory, String keyPrefix) {
+        if (!fromDirectory.isDirectory()) {
+            throw new IllegalArgumentException("'fromDirectory' must be a directory");
+        }
+        File uploadTargetDir = new File(uploadQueueDir, keyPrefix);
+        File[] uploadables = fromDirectory.listFiles();
         if (uploadables != null) {
             for (File uploadable : uploadables) {
                 try {
                     if (uploadable.isDirectory()) {
-                        File zipFile = new File(collectionWorkDir, uploadable.getName() + ".zip");
+                        File zipFile = new File(fromDirectory, uploadable.getName() + ".zip");
                         ZipUnzip.zip(uploadable, zipFile, true);
                         FileUtils.deleteDirectory(uploadable);
-                        FileUtils.moveFileToDirectory(zipFile, uploadQueueDeviceDir, true);
+                        FileUtils.moveFileToDirectory(zipFile, uploadTargetDir, true);
                     } else {
-                        FileUtils.moveFileToDirectory(uploadable, uploadQueueDeviceDir, true);
+                        FileUtils.moveFileToDirectory(uploadable, uploadTargetDir, true);
                     }
                 } catch (IOException e) {
                     // This really shouldn't happen. If it does, then what?
@@ -89,6 +87,9 @@ public class StatisticsUploader {
                     // such that it does, this code will at least work.
                     Thread.sleep(60000);
                 } else if (queue.size() > 0) {
+                    // Given an uploadQueue ~/Users/alice/Amplio/uploadQueue, and a file in that directory
+                    // collected-data/tbcd000c/20220110T085217.123Z.zip, upload the file to
+                    // s3://acm-stats/collected-data/tbcd000c/2022-110T085217.123Z.zip
                     File nextFile = queue.remove(0);
                     Path keyPath = Paths.get(nextFile.getAbsolutePath());
                     Path relativePath = uploadQueuePath.relativize(keyPath);
@@ -115,26 +116,30 @@ public class StatisticsUploader {
             tbLoader.updateUploadStatus(progress);
         }
     }
+
     UploadWorker uploadWorker;
 
 
     /**
      * Scans the upload queue for anything waiting to be uploaded.
+     *
      * @return a list of files in the upload queue.
      */
     private List<File> getUploadQueue() {
         List<File> result = new ArrayList<>();
-        StatisticsUploader.addFiles(uploadQueueCDDir, result, true);
+        StatisticsUploader.addFiles(uploadQueueDir, result, true);
         result.sort(Comparator.comparingLong(File::length));
         return result;
     }
 
     /**
      * Helper for recursively getting the files in a directory.
-     * @param file a file to be added, or a directory to be scanned recursively.
-     * @param list of files to be appended to.
+     *
+     * @param file        a file to be added, or a directory to be scanned recursively.
+     * @param list        of files to be appended to.
+     * @param removeEmpty if true, rmdir empty directories.
      */
-    static void addFiles(File file, List<File> list, boolean removeEmpty) {
+    private static void addFiles(File file, List<File> list, boolean removeEmpty) {
         if (!file.exists()) return;
         if (file.isDirectory()) {
             File[] dirList = file.listFiles();
@@ -162,6 +167,5 @@ public class StatisticsUploader {
             nBytes = files.stream().map(File::length).reduce(0L, Long::sum);
         }
     }
-
 
 }
